Add files via upload
I took Narjes file and merged in a bunch of changes.  I took out the date hard coding and made it a variable that you change in the second notebook slot
As a result, it will create a directory with the correct date range as a name, save the plot using that date range as the file name and will plug the correct date ranges into the plot title.
I also applied the seaborn library which gives the plots a more modern design.  

Finally, I integrated Annette's percentage change graphs so that all 5 time periods are plotted and saved.  

All the plots are in their respective directories.  Some are unusual looking because the vertical dotted lines skew them to the left or right.  I'm working on a way to test if that is a relevant feature in a plot, and if not, to skip it, but i wanted to get these up for review because it's a cosmetic defect, not a data defect, so I won't feel terrible if they're on a slide that I have to brief.